requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
mysql-connector-python==8.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
reportlab==4.0.8
